Change file extensions on description files
This commit changes the file extension from ".txt" to ".rst." The
purpose is to allow for better formatting in the README file. While the
descriptions will still appear in unprocessed plain text in the terminal
GUI and CLI, reStructuredText is human-readable enough that it shouldn't
be a problem. In the future, a reST to plain text converter can be used
for the terminal.