Update keybindings
With this commit, when the arrow keys are being used for moving the
focus in interactive menus, they can no longer move the focus outside
the menu. The purpose of this change is to allow the user to move
through the menu without accidentally selecting anything outside it.
Separate keybindings are used for moving focus outside the menu.