Fix issue creating predictable puzzles
On most puzzles, there would be two continuous columns which would only
overlap by one row. When this happens, the player only needs to look at
those two columns to solve the puzzle. Because this happened so
frequently, puzzles were easy to solve. This commit fixes the issue and
makes it a random, less frequent occurrence.